Rename sub-action modules to differentiate them
