+++ borg_the_docs/Dockerfile
@@ -8,6 +8,6 @@ RUN pip install github-action-utils
 
 ADD borg_the_docs.py /borg_the_docs.py
 ADD entrypoint.py /entrypoint.py
-ADD ntd2d_action /ntd2d_action
+ADD borg_the_docs_action /borg_the_docs_action_action
 
 ENTRYPOINT ["/entrypoint.py"]
