+++ update_pages/Dockerfile
@@ -24,6 +24,6 @@ RUN pip install github-action-utils
 RUN git config --global --add safe.directory /github/workspace
 
 ADD entrypoint.py /entrypoint.py
-ADD ntd2d_action /ntd2d_action
+ADD update_pages_action /update_pages_action
 
 ENTRYPOINT ["/entrypoint.py"]
