+++ update_pages/update_pages_action/files/templates/variant_item.html
