+++ update_pages/update_pages_action/files/templates/variants.html
